ci: fix errors that crop up with Wsign-compare

+++ cmidi2.h
@@ -712,14 +712,14 @@ static inline void cmidi2_ump_sysex_get_packet_of(uint64_t* result1, uint64_t* r
 
     enum cmidi2_sysex_status status;
     uint8_t size;
-    if (numBytes <= radix) {
+    if (numBytes <= (size_t)radix) {
         status =  CMIDI2_SYSEX_IN_ONE_UMP;
         size = numBytes; // single packet message
     } else if (index == 0) {
         status = CMIDI2_SYSEX_START;
         size = radix;
     } else {
-        uint8_t isEnd = index == cmidi2_ump_sysex_get_num_packets(numBytes, radix) - 1;
+        uint8_t isEnd = (size_t)index == cmidi2_ump_sysex_get_num_packets(numBytes, radix) - 1;
         if (isEnd) {
             size = numBytes % radix ? numBytes % radix : radix;
             status = CMIDI2_SYSEX_END;
@@ -941,7 +941,7 @@ static inline void cmidi2_ump_flex_data_get_packet_of(uint8_t group, uint8_t add
         format = CMIDI2_SYSEX_START;
         size = radix;
     } else {
-        uint8_t isEnd = currentPacket == cmidi2_ump_sysex_get_num_packets(numBytes, radix) - 1;
+        uint8_t isEnd = (size_t)currentPacket == cmidi2_ump_sysex_get_num_packets(numBytes, radix) - 1;
         if (isEnd) {
             size = numBytes % radix ? numBytes % radix : radix;
             format = CMIDI2_SYSEX_END;
